refactor: move reducers to game-logic
